Update text
Text elements update whenever a soup is cooked or swapped. There's still a bug with the active soups to fix where it doesn't show up

In scene: added UsesText as children under all icons; same UsesText as SpoonSelect. Also added SerializeField for each text object to be accessed in soupUI